hexalearn/requirements_light.txt
# ============================================================
# HexaLearn — AWS Free Tier build
# Đã bỏ: celery, redis, amqp, kombu, billiard, vine,
#         django_celery_results (thay bằng threading)
# ============================================================

annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
Django==6.0.3
django-cloudinary-storage==0.3.0
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
fugashi==1.5.2
google-auth==2.49.1
google-genai==1.71.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.1
psycopg2==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unidic-lite==1.0.8
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0